serial_size_t -> size_t

+++ examples/cifar10/test.cpp
@@ -46,9 +46,9 @@ void construct_net(N &nn) {
   using relu    = relu_layer;
   using softmax = softmax_layer;
 
-  const int n_fmaps  = 32;  ///< number of feature maps for upper layer
-  const int n_fmaps2 = 64;  ///< number of feature maps for lower layer
-  const int n_fc     = 64;  ///< number of hidden units in fully-connected layer
+  const size_t n_fmaps  = 32;  ///< number of feature maps for upper layer
+  const size_t n_fmaps2 = 64;  ///< number of feature maps for lower layer
+  const size_t n_fc     = 64;  ///< number of hidden units in fully-connected layer
 
   nn << conv(32, 32, 5, 3, n_fmaps, padding::same)        // C1
      << pool(32, 32, n_fmaps, 2)                          // P2
